Add instructions for Ubuntu 10.04.3 LTS
This is a terribly old version of Ubuntu (I'm just a guest on that
box), but it's the only version I have convenient access to.  Ubuntu
has been through a major UI overhaul since that version [1], so
screenshots from a more recent version would be good.  Still, some old
screenshots are probably better than no screenshots ;).

[1]: https://en.wikipedia.org/wiki/Unity_%28user_interface%29